chore: moved ast-serialization infra to validation
