+++ packages/debug/src/index.ts
@@ -10,9 +10,4 @@ export {
 
 export {
   Unparser,
-  Serializer,
-  Deserializer,
-  serializePrimitive,
-  serializePrimitives,
-  deserializePrimitive,
 } from './binding/unparser';
