+++ packages/validation/src/index.ts
@@ -9,3 +9,4 @@ export * from './subscribers/validation-errors-custom-attribute';
 export * from './subscribers/validation-result-presenter-service';
 export * from './rule-interfaces';
 export * from './serialization';
+export * from './ast-serialization';
